week1/cli_quiz/requirements.txt
pathlib>=1.0.1
PyYAML>=6.0.1
SQLAlchemy>=2.0.0 